feat(cpp,tests,docs,ci,compiler,scripts): M6 测试防线收尾 — 新增 59 个 C++ E2E 用例并纳入 CI/Shadow Verification
- 新增 59 个 C++ E2E 回归用例（全部通过），覆盖类/模板/容器/引用/unique_ptr/算法题

- 扩展 cide_e2e.rs 支持 C++ 用例加载、运行与 TEST_REPORT.md 统计

- 扩展 shadow_verify_cpp.py 自动扫描 cases/cpp/，76 MATCH / 5 个已记录 gap

- CI 新增 C++ Shadow Verification 步骤与报告 artifact 上传

- ci_three_tier_check.py 将 cide_e2e 与 CPP_FAILURES.md 纳入三层契约检查

- 更新 CPLUSPLUS_EXTENSION_PLAN.md、AGENTS.md、CHANGELOG.md、CPP_FAILURES.md

- 修复 29 个 clippy 警告，cargo clippy -- -D warnings 通过

+++ native/src/session.rs
@@ -280,26 +280,26 @@ impl MemoryState {
         let mut i = 0;
         while i < self.free_list.len() {
             let block = &self.free_list[i];
-            let block_end = block.addr as u32 + block.size as u32;
-            if (block.addr as u32) >= alloc_end || block_end <= addr {
+            let block_end = block.addr + block.size as u32;
+            if block.addr >= alloc_end || block_end <= addr {
                 i += 1;
                 continue;
             }
-            if (block.addr as u32) >= addr && block_end <= alloc_end {
+            if block.addr >= addr && block_end <= alloc_end {
                 // 完全被覆盖
                 self.free_list.remove(i);
-            } else if (block.addr as u32) < addr && block_end > alloc_end {
+            } else if block.addr < addr && block_end > alloc_end {
                 // 分配在块内部：拆分为前后两部分
                 let tail_size = block_end - alloc_end;
-                let head_size = addr - (block.addr as u32);
+                let head_size = addr - block.addr;
                 self.free_list[i].size = head_size as i32;
                 if tail_size > 0 {
                     self.free_list.push(FreeBlock { addr: alloc_end, size: tail_size as i32 });
                 }
                 i += 1;
-            } else if (block.addr as u32) < addr {
+            } else if block.addr < addr {
                 // 覆盖块的后部
-                self.free_list[i].size = (addr - block.addr as u32) as i32;
+                self.free_list[i].size = (addr - block.addr) as i32;
                 i += 1;
             } else {
                 // 覆盖块的前部
